requirements.txt
fastapi
uvicorn
Pillow
scikit-learn
torch
open-clip-torch
